backend/requirements.txt
# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# Audio Processing & AI
faster-whisper
pydub
audio-separator

# PyTorch (CUDA 12.x for cu128)
# --extra-index-url https://download.pytorch.org/whl/cu128
# torch
# torchvision
# torchaudio

# OCR (Optical Character Recognition)
# --extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cpu/
# paddlepaddle==3.2.0
# paddleocr

# Video & Online Streaming
ffmpeg-python
static_ffmpeg==3.0
requests==2.32.5
qrcode[pil]

#celery tasks
celery[redis]
python-multipart
fastapi
uvicorn